Refactor complete: Managers, AbstractWhale, and Fixes
